feat(frontend): revise ModalSheet and Sheet
