+++ services/frontend/app/components/ModalSheet/Modal/index.tsx
@@ -1,0 +1,41 @@
+import type { ModalSheetProps } from '~frontend/components/ModalSheet/types';
+import { Modal as UIModal } from '~frontend/components/Modal';
+import { getPassiveClose } from '~frontend/components/ModalSheet/utils';
+
+export function Modal({
+  onClose,
+  children,
+  title,
+  picture = null,
+  description,
+  actionVertical,
+  hasCloseButton,
+  severity,
+}: React.PropsWithChildren<ModalSheetProps>) {
+  const passiveClose = getPassiveClose(severity);
+
+  return (
+    <UIModal
+      onClose={onClose}
+      hasCloseButton={hasCloseButton}
+      clickOutsideToClose={!passiveClose}
+    >
+      <div className="mt-4 flex flex-col items-center gap-2">
+        {picture && <picture className="h-20 w-20">{picture}</picture>}
+        <header className="text-base font-bold color-zinc-700 md:text-lg">
+          {title}
+        </header>
+        <div className="mx-4 text-sm color-zinc-700 md:text-base">
+          {description}
+        </div>
+        <footer
+          className={`flex pt-3 px-4 mb-5 border-t-1 border-t-solid border-t-zinc-200 w-100% ${
+            actionVertical ? 'flex-col gap-1' : 'gap-2'
+          }`}
+        >
+          {children}
+        </footer>
+      </div>
+    </UIModal>
+  );
+}
