+++ services/frontend/app/components/ModalSheet/constants.ts
@@ -1,0 +1,1 @@
+export const SEVERITIES = ['info', 'warning', 'error'] as const;
