+++ services/frontend/app/components/ModalSheet/utils.ts
@@ -1,0 +1,2 @@
+import { Severity } from '~frontend/components/ModalSheet/types';
+export const getPassiveClose = (severity: Severity) => severity !== 'info';
